Struct nphysics3d::algebra::Velocity2

pub struct Velocity2<N: Real> {
    pub linear: Vector2<N>,
    pub angular: N,
}

A velocity structure combining both the linear angular velocities of a point.

Fields
linear: Vector2<N>

The linear velocity.

angular: N

The angular velocity.

Methods

impl<N: Real> Velocity2<N>

pub fn new(linear: Vector2<N>, angular: N) -> Self

Create velocity from its linear and angular parts.

pub fn new_with_vectors(linear: Vector2<N>, angular: Vector1<N>) -> Self

Create velocity from its linear and angular parts.

pub fn angular(w: N) -> Self

Create a purely angular velocity.

pub fn linear(vx: N, vy: N) -> Self

Create a purely linear velocity.

pub fn zero() -> Self

Create a zero velocity.

pub fn angular_vector(&self) -> Vector1<N>

The angular part of the velocity.

Important traits for &'a mut [u8]
impl<'a> Write for &'a mut [u8]impl<'a> Read for &'a [u8]

pub fn as_slice(&self) -> &[N]

This velocity seen as a slice.

The linear part is stored first.

Important traits for &'a mut [u8]
impl<'a> Write for &'a mut [u8]impl<'a> Read for &'a [u8]

pub fn as_mut_slice(&mut self) -> &mut [N]

This velocity seen as a mutable slice.

The linear part is stored first.

pub fn as_vector(&self) -> &Vector3<N>

This velocity seen as a vector.

The linear part is stored first.

pub fn as_vector_mut(&mut self) -> &mut Vector3<N>

This velocity seen as a mutable vector.

The linear part is stored first.

pub fn from_vector<S: Storage<N, U3>>(data: &Vector<N, U3, S>) -> Self

Create a velocity from a vector.

The linear part of the velocity is expected to be first inside of the input vector.

pub fn from_slice(data: &[N]) -> Self

Create a velocity from a slice.

The linear part of the velocity is expected to be first inside of the input slice.

pub fn shift(&self, shift: &Vector2<N>) -> Self

Compute the velocity of a point that is located at the coordinates shift relative to the point having self as velocity.

pub fn rotated(&self, rot: &Rotation2<N>) -> Self

Rotate each component of self by rot.

pub fn transformed(&self, iso: &Isometry2<N>) -> Self

Transform each component of self by iso.