Struct nphysics3d::counters::Counters

pub struct Counters { /* fields omitted */ }

Aggregation of all the performances counters tracked by nphysics.

Methods

impl Counters

pub fn new(enabled: bool) -> Self

Create a new set of counters initialized to wero.

pub fn enable(&mut self)

Enable all the counters.

pub fn enabled(&self) -> bool

Return true if the counters are enabled.

pub fn disable(&mut self)

Disable all the counters.

pub fn step_started(&mut self)

Notify that the time-step has started.

pub fn step_completed(&mut self)

Notfy that the time-step has finished.

pub fn step_time(&self) -> f64

Total time spent for one step of the physics engine.

pub fn set_nconstraints(&mut self, n: usize)

Set the number of constraints generated.

pub fn set_ncontact_pairs(&mut self, n: usize)

Set the number of contact pairs generated.

impl Counters

pub fn update_started(&mut self)

Start this timer.

pub fn update_completed(&mut self)

Stop this timer.

pub fn update_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn collision_detection_started(&mut self)

Start this timer.

pub fn collision_detection_completed(&mut self)

Stop this timer.

pub fn collision_detection_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn island_construction_started(&mut self)

Start this timer.

pub fn island_construction_completed(&mut self)

Stop this timer.

pub fn island_construction_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn solver_started(&mut self)

Start this timer.

pub fn solver_completed(&mut self)

Stop this timer.

pub fn solver_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn assembly_started(&mut self)

Start this timer.

pub fn assembly_completed(&mut self)

Stop this timer.

pub fn assembly_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn resolution_started(&mut self)

Start this timer.

pub fn resolution_completed(&mut self)

Stop this timer.

pub fn resolution_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn position_update_started(&mut self)

Start this timer.

pub fn position_update_completed(&mut self)

Stop this timer.

pub fn positon_update_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn broad_phase_started(&mut self)

Start this timer.

pub fn broad_phase_completed(&mut self)

Stop this timer.

pub fn broad_phase_time(&self) -> f64

Gets the time elapsed for this timer.

impl Counters

pub fn narrow_phase_started(&mut self)

Start this timer.

pub fn narrow_phase_completed(&mut self)

Stop this timer.

pub fn narrow_phase_time(&self) -> f64

Gets the time elapsed for this timer.