Struct nphysics3d::force_generator::ConstantAcceleration

pub struct ConstantAcceleration<N: Real> { /* fields omitted */ }

Force generator adding a constant acceleration at the center of mass of a set of body parts.

Methods

impl<N: Real> ConstantAcceleration<N>

pub fn new(linear_acc: Vector<N>, angular_acc: Vector<N>) -> Self

Adds a new constant acceleration generator.

The acceleration is expressed in world-space.

pub fn add_body_part(&mut self, body: BodyHandle)

Add a body part to be affected by this force generator.