Module nphysics3d::joint

Joints using the reduced-coordinates formalism or using constraints.

Structs

BallConstraint
A constraint that removes all relative linear motion between two body parts.

BallJoint
A joint that allows only all rotational degrees of freedom between two multibody links.

CartesianConstraint
A constraint that removes all relative angular motion between two body parts.

CartesianJoint
A joint that allows only all the translational degrees of freedom between two multibody links.

CylindricalConstraint
A constraint that removes all degrees of freedom (of one body part relative to a second one) except one translation along an axis and one rotation along the same axis.

CylindricalJoint
A joint that allows one translational and one rotational degrees of freedom along a single axis.

FixedConstraint
A constraint that removes all degrees of freedom between two body parts.

FixedJoint
A joint that does not allow any relative degrees of freedom.

FreeJoint
A joint that allows all the relative degrees of freedom between two multibody links.

HelicalJoint
A joint that allows one degree of freedom between two multibody links.

JointMotor
Description of a motor applied to a joint.

MouseConstraint
A spring-like constraint to be used to drag a body part with the mouse.

PinSlotConstraint
A constraint that removes two translational and two rotational degrees of freedoms.

PinSlotJoint
A joint that allows one translational and one rotational degrees of freedom.

PlanarConstraint
A constraint that removes one relative translational degree of freedom, and all but one rotational degrees of freedom.

PlanarJoint
A joint that allows 1 rotational and 2 translational degrees of freedom.

PrismaticConstraint
A constraint that remove all be one translational degrees of freedom.

PrismaticJoint
A unit joint that allows only one translational degree on freedom.

RectangularConstraint
A constraint that remove all relative rotations and one relative translation between two body parts.

RectangularJoint
A joint that allows two translational degrees of freedom.

RevoluteConstraint
A constraint that removes all relative motions except one rotation between two body parts.

RevoluteJoint
A unit joint that allows only one relative rotational degree of freedom between two multibody links.

UniversalConstraint
A constraint that removes all but two relative rotations along distinct axii.

UniversalJoint
A joint that allows only two relative rotations between two multibody links.

Traits

Joint
Trait implemented by all joints following the reduced-coordinate formation.

JointConstraint
Trait implemented by joint that operate by generating constraints to restrict the relative motion of two body parts.

UnitJoint
Trait implemented by joints using the reduced-coordinates approach and allowing only one degree of freedom.

Functions

unit_joint_num_velocity_constraints
Computes the maximum number of velocity constraints to be applied by the given unit joint.

unit_joint_position_constraint
Initializes and generate the position constraints applicable to the multibody links attached to this joint.

unit_joint_velocity_constraints
Initializes and generate the velocity constraints applicable to the multibody links attached to this joint.

Type Definitions

ConstraintHandle
The handle of a consraint.