Struct nphysics3d::joint::FreeJoint

pub struct FreeJoint<N: Real> { /* fields omitted */ }

A joint that allows all the relative degrees of freedom between two multibody links.

This joint can only be added between a Ground body (as parent) and any other body.

Methods

impl<N: Real> FreeJoint<N>

pub fn new(position: Isometry<N>) -> Self

Creates a free joint with the given initial position of the descendent, relative to the ground.