Struct nphysics3d::joint::PrismaticJoint

pub struct PrismaticJoint<N: Real> { /* fields omitted */ }

A unit joint that allows only one translational degree on freedom.

Methods

impl<N: Real> PrismaticJoint<N>

pub fn new(axis: Unit<Vector<N>>, offset: N) -> Self

Create a new prismatic joint where the allowed traslation is defined along the provided axis.

The axis is expressed in the local coordinate system of the two multibody links attached to this joint.

pub fn offset(&self) -> N

The relative displacement of the attached multibody links along the joint axis.

pub fn translation(&self) -> Translation<N>

The relative translation of the attached multibody links along the joint axis.

pub fn min_offset(&self) -> Option<N>

The lower limit of the relative displacement of the attached multibody links along the joint axis.

pub fn max_offset(&self) -> Option<N>

The upper limit of the relative displacement of the attached multibody links along the joint axis.

pub fn disable_min_offset(&mut self)

Disable the lower limit of the relative displacement of the attached multibody links along the joint axis.

pub fn disable_max_offset(&mut self)

Disable the upper limit of the relative displacement of the attached multibody links along the joint axis.

pub fn enable_min_offset(&mut self, limit: N)

Set the lower limit of the relative displacement of the attached multibody links along the joint axis.

pub fn enable_max_offset(&mut self, limit: N)

Set the upper limit of the relative displacement of the attached multibody links along the joint axis.

pub fn is_linear_motor_enabled(&self) -> bool

Returns true if the joint motor is enabled.

pub fn enable_linear_motor(&mut self)

Enable the joint motor.

pub fn disable_linear_motor(&mut self)

Disable the joint motor.

pub fn desired_linear_motor_velocity(&self) -> N

The desired relative velocity to be enforced by the joint motor.

pub fn set_desired_linear_motor_velocity(&mut self, vel: N)

Set the desired relative velocity to be enforced by the joint motor.

pub fn max_linear_motor_force(&self) -> N

The maximum force that can be output by the joint motor.

pub fn set_max_linear_motor_force(&mut self, force: N)

Set the maximum force that can be output by the joint motor.