Struct nphysics3d::joint::RectangularJoint

pub struct RectangularJoint<N: Real> { /* fields omitted */ }

A joint that allows two translational degrees of freedom.

Methods

impl<N: Real> RectangularJoint<N>

pub fn new(
axis1: Unit<Vector3<N>>,
axis2: Unit<Vector3<N>>,
offset1: N,
offset2: N
) -> Self

Creates a new rectangular joint allowing relative translations anlon the two provided axii.Isometry3

Both axii are expressed in the local coordinate frame on the attached multibody links.

impl<N: Real> RectangularJoint<N>

pub fn min_offset_1(&self) -> Option<N>

The lower limit of the relative translational displacement of the attached multibody links along the joint axis.

pub fn max_offset_1(&self) -> Option<N>

The upper limit of the relative translational displacement of the attached multibody links along the joint axis.

pub fn disable_min_offset_1(&mut self)

Disable the lower limit of the relative translational displacement of the attached multibody links along the joint axis.

pub fn disable_max_offset_1(&mut self)

Disable the upper limit of the relative translational displacement of the attached multibody links along the joint axis.

pub fn enable_min_offset_1(&mut self, limit: N)

Set the lower limit of the relative translational displacement of the attached multibody links along the joint axis.

pub fn enable_max_offset_1(&mut self, limit: N)

Set the upper limit of the relative translational displacement of the attached multibody links along the joint axis.

pub fn is_linear_motor_enabled_1(&self) -> bool

Returns true if the joint translational motor is enabled.

pub fn enable_linear_motor_1(&mut self)

Enable the joint translational motor.

pub fn disable_linear_motor_1(&mut self)

Disable the joint translational motor.

pub fn desired_linear_motor_velocity_1(&self) -> N

The desired relative translational velocity to be enforced by the joint motor.

pub fn set_desired_linear_motor_velocity_1(&mut self, vel: N)

Set the desired relative translational velocity to be enforced by the joint motor.

pub fn max_linear_motor_force_1(&self) -> N

The maximum force that can be output by the joint translational motor.

pub fn set_max_linear_motor_force_1(&mut self, force: N)

Set the maximum force that can be output by the joint translational motor.

impl<N: Real> RectangularJoint<N>

pub fn min_offset_2(&self) -> Option<N>

The lower limit of the relative translational displacement of the attached multibody links along the joint axis.

pub fn max_offset_2(&self) -> Option<N>

The upper limit of the relative translational displacement of the attached multibody links along the joint axis.

pub fn disable_min_offset_2(&mut self)

Disable the lower limit of the relative translational displacement of the attached multibody links along the joint axis.

pub fn disable_max_offset_2(&mut self)

Disable the upper limit of the relative translational displacement of the attached multibody links along the joint axis.

pub fn enable_min_offset_2(&mut self, limit: N)

Set the lower limit of the relative translational displacement of the attached multibody links along the joint axis.

pub fn enable_max_offset_2(&mut self, limit: N)

Set the upper limit of the relative translational displacement of the attached multibody links along the joint axis.

pub fn is_linear_motor_enabled_2(&self) -> bool

Returns true if the joint translational motor is enabled.

pub fn enable_linear_motor_2(&mut self)

Enable the joint translational motor.

pub fn disable_linear_motor_2(&mut self)

Disable the joint translational motor.

pub fn desired_linear_motor_velocity_2(&self) -> N

The desired relative translational velocity to be enforced by the joint motor.

pub fn set_desired_linear_motor_velocity_2(&mut self, vel: N)

Set the desired relative translational velocity to be enforced by the joint motor.

pub fn max_linear_motor_force2(&self) -> N

The maximum force that can be output by the joint translational motor.

pub fn set_max_linear_motor_force_2(&mut self, force: N)

Set the maximum force that can be output by the joint translational motor.