Enum nphysics3d::object::BodyMut

pub enum BodyMut<'a, N: Real> {
    RigidBody(&'a mut RigidBody<N>),
    Multibody(&'a mut Multibody<N>),
    Ground(&'a mut Ground<N>),
}

A mutable body contained by the physics world.

Variants
RigidBody(&'a mut RigidBody<N>)

A rigid body.

Multibody(&'a mut Multibody<N>)

A multibody.

Ground(&'a mut Ground<N>)

The ground.

Methods

impl<'a, N: Real> BodyMut<'a, N>

pub fn is_ground(&self) -> bool

Checks if this body identifies the ground.

pub fn status_dependent_ndofs(&self) -> usize

The number of degrees of freedom (DOF) of this body, taking its status into account.

In particular, this returns 0 for any body with a status different than BodyStatus::Dynamic.

pub fn apply_displacement(&mut self, disp: &[N])

Applies a displacement to all the degrees of freedom of this body.

impl<'a, N: Real> BodyMut<'a, N>

pub fn handle(&self) -> BodyHandle

The handle of this body.

impl<'a, N: Real> BodyMut<'a, N>

pub fn status(&self) -> BodyStatus

The status of this body.

impl<'a, N: Real> BodyMut<'a, N>

pub fn activation_status(&self) -> &ActivationStatus<N>

Informations regarding activation and deactivation (sleeping) of this body.

impl<'a, N: Real> BodyMut<'a, N>

pub fn is_active(&self) -> bool

Check if this body is active.

impl<'a, N: Real> BodyMut<'a, N>

pub fn is_kinematic(&self) -> bool

Whether or not the status of this body is kinematic.

impl<'a, N: Real> BodyMut<'a, N>

pub fn is_static(&self) -> bool

Whether or not the status of this body is static.

impl<'a, N: Real> BodyMut<'a, N>

pub fn ndofs(&self) -> usize

The number of degrees of freedom of this body.

impl<'a, N: Real> BodyMut<'a, N>

pub fn generalized_acceleration(&self) -> DVectorSlice<N>

The generalized accelerations at each degree of freedom of this body.

impl<'a, N: Real> BodyMut<'a, N>

pub fn generalized_velocity(&self) -> DVectorSlice<N>

The generalized velocities of this body.

impl<'a, N: Real> BodyMut<'a, N>

pub fn companion_id(&self) -> usize

The companion ID of this body.

impl<'a, N: Real> BodyMut<'a, N>

pub fn set_companion_id(&mut self, id: usize)

Set the companion ID of this body (may be reinitialized by nphysics).

impl<'a, N: Real> BodyMut<'a, N>

pub fn generalized_velocity_mut(&mut self) -> DVectorSliceMut<N>

The mutable generalized velocities of this body.

impl<'a, N: Real> BodyMut<'a, N>

pub fn integrate(&mut self, params: &IntegrationParameters<N>)

Integrate the position of this body.

impl<'a, N: Real> BodyMut<'a, N>

pub fn activate(&mut self)

Force the activation of this body.

impl<'a, N: Real> BodyMut<'a, N>

pub fn activate_with_energy(&mut self, energy: N)

Force the activation of this body with the given level of energy.

impl<'a, N: Real> BodyMut<'a, N>

pub fn deactivate(&mut self)

Put this body to sleep.