Enum nphysics3d::object::BodyPart

pub enum BodyPart<'a, N: Real> {
    RigidBody(&'a RigidBody<N>),
    MultibodyLink(MultibodyLinkRef<'a, N>),
    Ground(&'a Ground<N>),
}

A mutable body part contained by the physics world.

Variants
RigidBody(&'a RigidBody<N>)

A rigid body.

MultibodyLink(MultibodyLinkRef<'a, N>)

A link of a multibody.

Ground(&'a Ground<N>)

The ground.

Methods

impl<'a, N: Real> BodyPart<'a, N>

pub fn is_ground(&self) -> bool

Returns true if this body is the ground.

pub fn handle(&self) -> BodyHandle

The handle of this body.

pub fn parent_ndofs(&self) -> usize

The number of degrees of freedom of the body containing this body part.

pub fn parent_companion_id(&self) -> usize

The companion ID of the parent of this body part.

pub fn is_active(&self) -> bool

Check if this body part is active.

pub fn status_dependent_parent_ndofs(&self) -> usize

The number of degrees of freedom (DOF) of the body containing this body part, taking its status into account.

In particular, this returns 0 for any body with a status different than BodyStatus::Dynamic.

pub fn parent_generalized_velocity(&self) -> DVectorSlice<N>

The generalized velocities of the body containing this body part.

pub fn center_of_mass(&self) -> Point<N>

The center of mass of this body part.

pub fn position(&self) -> Isometry<N>

The position of this body part wrt. the ground.

pub fn velocity(&self) -> Velocity<N>

The velocity of this body part.

pub fn inertia(&self) -> Inertia<N>

The world-space inertia of this body part.

pub fn local_inertia(&self) -> Inertia<N>

The local-space inertia of this body part.

pub fn status_dependent_velocity(&self) -> Velocity<N>

The velocity of this body part, taking it status into account.

This always return zero if the body part is static.

pub fn body_jacobian_mul_force(&self, force: &Force<N>, out: &mut [N])

Convert a force applied to the center of mass of this body part into generalized force.

pub fn inv_mass_mul_generalized_forces(&self, out: &mut [N])

Convert generalized forces applied to this body part into generalized accelerations.

pub fn inv_mass_mul_force(&self, force: &Force<N>, out: &mut [N])

Convert a force applied to this body part's center of mass into generalized accelerations.