Enum nphysics3d::object::BodyPartMut

pub enum BodyPartMut<'a, N: Real> {
    RigidBody(&'a mut RigidBody<N>),
    MultibodyLink(MultibodyLinkMut<'a, N>),
    Ground(&'a mut Ground<N>),
}

A mutable body part contained by the physics world.

Variants
RigidBody(&'a mut RigidBody<N>)

A rigid body.

MultibodyLink(MultibodyLinkMut<'a, N>)

A link of a multibody.

Ground(&'a mut Ground<N>)

The ground.

Methods

impl<'a, N: Real> BodyPartMut<'a, N>

pub fn as_ref<'b>(&'b self) -> BodyPart<'b, N>

Retrieves an immutable reference to this body part.

pub fn apply_force(&mut self, force: &Force<N>)

Apply a force to this body part at the next frame.