Struct nphysics3d::object::Ground

pub struct Ground<N: Real> { /* fields omitted */ }

A singleton representing the ground.

Most of its methods are useless but provided anyway to be similar to the other bodies.

Methods

impl<N: Real> Ground<N>

pub fn handle(&self) -> BodyHandle

The handle of the ground.

pub fn activation_status(&self) -> &ActivationStatus<N>

An inactive status.

pub fn activate(&mut self)

Does nothing.

pub fn activate_with_energy(&mut self, energy: N)

Does nothing.

pub fn deactivate(&mut self)

Does nothing.

pub fn is_active(&self) -> bool

Returns false.

pub fn is_dynamic(&self) -> bool

Returns false.

pub fn is_kinematic(&self) -> bool

Returns false.

pub fn is_static(&self) -> bool

Returns true.

pub fn status(&self) -> BodyStatus

Returns BodyStatus::Static.

pub fn companion_id(&self) -> usize

The companion ID of the ground.

May change at each step of the physics engine.

pub fn set_companion_id(&mut self, id: usize)

Sets the companion ID of the ground.

This may be reset by the physics engine at each step.

pub fn ndofs(&self) -> usize

Returns 0.

pub fn generalized_velocity(&self) -> DVectorSlice<N>

Returns an empty slice.

pub fn generalized_velocity_mut(&mut self) -> DVectorSliceMut<N>

Returns an empty slice.

pub fn generalized_acceleration(&self) -> DVectorSlice<N>

Returns an empty slice.

pub fn integrate(&mut self, _: &IntegrationParameters<N>)

Does nothing.

pub fn center_of_mass(&self) -> Point<N>

Returns the origin.

pub fn position(&self) -> Isometry<N>

Returns the identity.

pub fn velocity(&self) -> Velocity<N>

Returns zero.

pub fn inertia(&self) -> Inertia<N>

Returns zero.

pub fn local_inertia(&self) -> Inertia<N>

Returns zero.

pub fn body_jacobian_mul_force(&self, _: &Force<N>, _: &mut [N])

Does nothing.

pub fn inv_mass_mul_generalized_forces(&self, _: &mut [N])

Does nothing.

pub fn inv_mass_mul_force(&self, _: &Force<N>, _: &mut [N])

Does nothing.

pub fn apply_force(&mut self, _: &Force<N>)

Does nothing.