Struct nphysics3d::object::MultibodyLinkMut

pub struct MultibodyLinkMut<'a, N: Real> { /* fields omitted */ }

Mutable reference to a multibody link.

Methods

impl<'a, N: Real> MultibodyLinkMut<'a, N>

pub fn new(id: MultibodyLinkId, multibody: &'a mut Multibody<N>) -> Self

Creates a new mutable reference of a multibody link.

pub fn handle(&self) -> BodyHandle

The handle of this link.

pub fn as_ref<'b>(&'b self) -> MultibodyLinkRef<'b, N>

Retrieve an immutable reference to this link.

pub fn joint_mut(&mut self) -> &mut Joint<N>

Mutable reference to the joint attaching this link to its parent.

pub fn apply_force(&mut self, force: &Force<N>)

Apply a force to this link.

pub fn joint_velocity_mut(&mut self) -> DVectorSliceMut<N>

Retriev the mutable generalized velocities of this link.