Struct nphysics3d::object::MultibodyLinkRef

pub struct MultibodyLinkRef<'a, N: Real> { /* fields omitted */ }

A reference to a multibody link.

Methods

impl<'a, N: Real> MultibodyLinkRef<'a, N>

pub fn new(id: MultibodyLinkId, multibody: &'a Multibody<N>) -> Self

Create a new reference to a multibody link.

pub fn handle(&self) -> BodyHandle

The handle of this link.

pub fn joint(&self) -> &Joint<N>

Reference to the joint attaching this link to its parent.

pub fn center_of_mass(&self) -> Point<N>

Center of mass of this link wrt. the ground.

pub fn velocity(&self) -> &Velocity<N>

Velocity at the center of mass of this link.

pub fn joint_velocity(&self) -> DVectorSlice<N>

Generalized velocities of this link.

pub fn id(&self) -> MultibodyLinkId

The identifier of this link.

pub fn assembly_id(&self) -> usize

The dynamic assembly identifier of this link.

pub fn impulse_id(&self) -> usize

The impulse cache identifier of this link.

pub fn multibody(&self) -> &Multibody<N>

Reference to the multibody containing this link.

pub fn inv_mass_mul_unit_joint_force(
&self,
dof_id: usize,
force: N,
out: &mut [N]
)

Convert a generalized force applied to this link degrees of freedom into generalized accelerations.

The joint attaching this link to its parent is assumed to be a unit joint.

pub fn inv_mass_mul_joint_force(&self, force: DVectorSlice<N>, out: &mut [N])

Convert a generalized force applied to this link degrees of freedom into generalized accelerations.

pub fn position(&self) -> Isometry<N>

The position of the center of mass of this link wrt. the ground.

pub fn body_jacobian_mul_force(&self, force: &Force<N>, out: &mut [N])

Convert a force applied to this link center of mass into generalized force.

pub fn inv_mass_mul_generalized_forces(&self, out: &mut [N])

Convert generalized forces applied to the multibody containing this link into generalized accelerations.

pub fn inv_mass_mul_force(&self, force: &Force<N>, out: &mut [N])

Convert a force applied to this link's center of mass into generalized accelerations.

pub fn local_inertia(&self) -> &Inertia<N>

The local inertia of this link.

pub fn inertia(&self) -> &Inertia<N>

The world-space inertia of this link.