Struct nphysics3d::object::RigidBodyCollisionGroups

pub struct RigidBodyCollisionGroups {
    // some fields omitted
}

Groups of collision used to filter which object collide with which other one. nphysics use a specific group for its own purposes (i.e. the group of static objects). The group 29 is reserved and you cannot use it.

Methods

impl RigidBodyCollisionGroups

fn new_dynamic() -> RigidBodyCollisionGroups

Creates a new RigidBodyCollisionGroups that enables collisions with everything except self-collision.

fn new_static() -> RigidBodyCollisionGroups

Creates a new RigidBodyCollisionGroups that enables collisions with every user-defined groups. Static objects and sensors are blacklisted by default and self-collision disabled.

fn is_dynamic(&self) -> bool

Returns true if this object is not part of the static group.

fn is_static(&self) -> bool

Returns true if this object is part of the static group.

impl RigidBodyCollisionGroups

fn max_group_id() -> usize

The maximum allowed group identifier.

fn as_collision_groups(&self) -> &CollisionGroups

Return the internal, ncollide-compatible, CollisionGroups

fn modify_membership(&mut self, group_id: usize, add: bool)

Adds or removes this entity from the given group.

fn modify_whitelist(&mut self, group_id: usize, add: bool)

Adds or removes the given group from this entity whitelist.

fn modify_blacklist(&mut self, group_id: usize, add: bool)

Adds or removes this entity from the given group.

fn set_membership(&mut self, groups: &[usize])

Make this object member of the given groups only.

fn set_whitelist(&mut self, groups: &[usize])

Whitelists the given groups only (others will be un-whitelisted).

fn set_blacklist(&mut self, groups: &[usize])

Blacklists the given groups only (others will be un-blacklisted).

fn copy_membership(&mut self, other: &RigidBodyCollisionGroups)

Copies the membership of another collision groups.

fn copy_whitelist(&mut self, other: &RigidBodyCollisionGroups)

Copies the whitelist of another collision groups.

fn copy_blacklist(&mut self, other: &RigidBodyCollisionGroups)

Copies the blacklist of another collision groups.

fn enable_interaction_with_static(&mut self)

Un-blacklists static objects.

fn disable_interaction_with_static(&mut self)

Blacklists any static object.

fn enable_interaction_with_sensors(&mut self)

Un-blacklists sensors.

fn disable_interaction_with_sensors(&mut self)

Blacklists sensors.

fn enable_self_interaction(&mut self)

Enables self interaction detection.

fn disable_self_interaction(&mut self)

Disables self interaction detection.

fn is_member_of(&self, group_id: usize) -> bool

Tests if this entity is part of the given group.

fn is_group_whitelisted(&self, group_id: usize) -> bool

Tests if the given group is whitelisted.

fn is_group_blacklisted(&self, group_id: usize) -> bool

Tests if the given group is blacklisted.

fn can_interact_with(&self, group_id: usize) -> bool

Tests whether collisions with a given group is possible.

Collision is possible if group_id is whitelisted but not blacklisted.

fn can_interact_with_groups(&self, other: &CollisionGroups) -> bool

Tests whether two collision groups have at least one group in common.

fn can_interact_with_self(&self) -> bool

Tests whether self-collision is enabled.