Struct nphysics3d::resolution::VelocityConstraint

pub struct VelocityConstraint<N: Scalar> {
    pub normal: Vector<N>,
    pub weighted_normal1: Vector<N>,
    pub weighted_normal2: Vector<N>,
    pub rot_axis1: Orientation<N>,
    pub weighted_rot_axis1: Orientation<N>,
    pub rot_axis2: Orientation<N>,
    pub weighted_rot_axis2: Orientation<N>,
    pub inv_projected_mass: N,
    pub impulse: N,
    pub lobound: N,
    pub hibound: N,
    pub objective: N,
    pub id1: isize,
    pub id2: isize,
    pub friction_limit_id: usize,
    pub friction_coeff: N,
}

A constraint of velocity at a point of contact.

Fields
normal: Vector<N>

The contact normal.

weighted_normal1: Vector<N>

The contact normal multiplied by the first body's inverse mass.

weighted_normal2: Vector<N>

The contact normal multiplied by the second body's inverse mass.

rot_axis1: Orientation<N>

The first body rotation axis.

weighted_rot_axis1: Orientation<N>

The first body rotation axis multiplied by its inverse inertia.

rot_axis2: Orientation<N>

The second body rotation axis.

weighted_rot_axis2: Orientation<N>

The second body rotation axis multiplied by its inverse inertia.

inv_projected_mass: N

The inverse of the sum of linear and angular inertia of both bodies.

impulse: N

The total impulse applied.

lobound: N

The lower bound of the impulse.

hibound: N

The upper bound of the impulse.

objective: N

The target delta velocity.

id1: isize

The id of the first body.

id2: isize

The id of the second body.

friction_limit_id: usize

The id of the friction constraint.

friction_coeff: N

The friction coefficient on this contact.

Methods

impl<N: Scalar> VelocityConstraint<N>

fn new() -> VelocityConstraint<N>

Creates a new velocity constraint with all terms initialized to zero.