Module nphysics3d::solver

Constraint solver.

Re-exports

pub use self::helper::ForceDirection;

Modules

helper
Utilities for computing velocity and position constraints.

Structs

BilateralConstraint
A bilateral (equality) constraint between two bodies.

BilateralGroundConstraint
A bilateral (equality) constraint between a dynamic body and one without any degrees of freedom.

ConstraintGeometry
Logical information of the geometry of a constraint.

ConstraintSet
A set of all velocity constraints and non-linear position-based constraints.

GenericNonlinearConstraint
A generic non-linear position constraint.

ImpulseCache
A cache for impulses.

IntegrationParameters
Parameters for a time-step of the physics engine.

MoreauJeanSolver
Moreau-Jean time-stepping scheme.

MultibodyJointLimitsNonlinearConstraintGenerator
A non-linear position constraint generator to enforce multibody joint limits.

NonlinearSORProx
Non-linear position-based consraint solver using the SOR-Prox approach.

NonlinearUnilateralConstraint
A non-linear position-based non-penetration constraint.

SORProx
A SOR-Prox velocity-based constraints solver.

SignoriniCoulombPyramidModel
A contact model generating one non-penetration constraint and two friction constraints per contact.

SignoriniModel
A contact model generating one non-penetration constraint per contact.

UnilateralConstraint
A unilateral (inequality) consraint.

UnilateralGroundConstraint
A unilateral (inequality) constraint between a dynamic body and one without any degrees of freedom.

Enums

ImpulseLimits
Limits of impulse applicable by a bilateral constraint.

Traits

ContactModel
The modeling of a contact.

NonlinearConstraintGenerator
Implemented by structures that generate non-linear constraints.