Module nphysics3d::solver::helper

Utilities for computing velocity and position constraints.

Enums

ForceDirection
The direction of a force.

Functions

align_axis
Generate position constraints to moved the body parts such that the given axis will become aligned.

cancel_relative_angular_velocity
Generate velocity constraints to cancel the relative angular velocity of two bodies.

cancel_relative_angular_velocity_wrt_axis
Generate velocity constraints to cancel the relative angular velocity of two bodies wrt. the given axis.

cancel_relative_linear_velocity
Generates velocity constraints to cancel the relative linear velocity of two body parts.

cancel_relative_linear_velocity_wrt_axis
Generates velocity constraints to cancel the relative linear velocity of two body parts wrt the given axis.

cancel_relative_rotation
Generate position constraints to cancel the relative rotation of two bodies.

cancel_relative_translation
Generate position constraints to cancel the relative translation of two bodies.

cancel_relative_translation_wrt_axis
Generate position constraints to cancel the relative translation of two bodies wrt the given axis.

constraint_pair_geometry
Fills all the jacobians (and the jacobians multiplied by the invers augmented mass matricxs) for a constraint applying a force at the points center1, center2 and the direction dir.

constraint_pair_velocity
Compute the generalized relative velocity for at the given points and along a given direction.

constraints_are_ground_constraints
Test sif a constraint between the two given bodies should be a ground constraint (a constraint between a dynamic body and one without any degree of freedom).

fill_constraint_geometry
Fills all the jacobians (and the jacobians multiplied by the invers augmented mass matricxs) for a constraint applying a force at the point center (relative to the body part's center of mass) and the direction dir.

project_anchor_to_axis
Generate position constraints to project anchor2 into the axis with direction axis1 and passing through the anchor1.

restore_angle_between_axis
Generate position constraints to ensure the two given axis are seperated by the given angle.

restrict_relative_angular_velocity_to_axis
Generate velocity constraints to cancel the relative angular velocity of two bodies along all axis except the one provided.

restrict_relative_linear_velocity_to_axis
Generate velocity constraints to cancel the relative linear velocity of two bodies along all axis except the one provided.