Function nphysics3d::solver::helper::fill_constraint_geometry

pub fn fill_constraint_geometry<N: Real>(
    body: &BodyPart<N>, 
    ndofs: usize, 
    center: &Point<N>, 
    dir: &ForceDirection<N>, 
    j_id: usize, 
    wj_id: usize, 
    jacobians: &mut [N], 
    inv_r: &mut N
)

Fills all the jacobians (and the jacobians multiplied by the invers augmented mass matricxs) for a constraint applying a force at the point center (relative to the body part's center of mass) and the direction dir.

If the force is a torque, it is applied at the center of mass of the body part.