Struct nphysics3d::solver::UnilateralConstraint

pub struct UnilateralConstraint<N: Real> {
    pub impulse: N,
    pub r: N,
    pub rhs: N,
    pub impulse_id: usize,
    pub assembly_id1: usize,
    pub assembly_id2: usize,
    pub j_id1: usize,
    pub j_id2: usize,
    pub wj_id1: usize,
    pub wj_id2: usize,
    pub ndofs1: usize,
    pub ndofs2: usize,
}

A unilateral (inequality) consraint.

Fields
impulse: N

The impulse applied by this constraint.

r: N

The scaling parameter of the SOR-prox method.

rhs: N

The target velocity change this constraint must apply.

impulse_id: usize

The index of the impulse used for its storage in an impuse cache.

assembly_id1: usize

The assembly index of the first body.

assembly_id2: usize

The assembly index of the second body.

j_id1: usize

Index of the first entry of the jacobian of the constraint affecting the first body.

j_id2: usize

Index of the first entry of the constraint jacobian multiplied by the inverse mass of the second body.

wj_id1: usize

Index of the first entry of the constraint jacobian multiplied by the inverse mass of the first body.

wj_id2: usize

Index of the first entry of the constraint jacobian multiplied by the inverse mass of the second body.

ndofs1: usize

Number of degree of freedom of the first body.

ndofs2: usize

Number of degree of freedom of the second body.

Methods

impl<N: Real> UnilateralConstraint<N>

pub fn new(
geom: ConstraintGeometry<N>,
assembly_id1: usize,
assembly_id2: usize,
rhs: N,
impulse: N,
impulse_id: usize
) -> Self

Create a new unilateral constraint.