Module nphysics3d::volumetric

Volume and inertia tensor computation.

Traits

InertiaTensor
Trait implemented by inertia tensors.

Volumetric
Trait implemented by objects which have a mass, a center of mass, and an inertia tensor.

Functions

ball_area
The area of a ball.

ball_center_of_mass
The center of mass of a ball.

ball_unit_angular_inertia
The unit angular inertia of a ball.

ball_volume
The volume of a ball.

cone_area
The area of a cone.

cone_center_of_mass
The center of mass of a cone.

cone_unit_angular_inertia
The unit angular inertia of a cone.

cone_volume
The volume of a cone.

convex_hull_area
The area of a convex hull.

convex_hull_center_of_mass
The center of mass of the convex hull of a set of points.

convex_hull_unit_angular_inertia
The angular inertia of the convex hull of a set of points.

convex_hull_volume
The volume of the convex hull of a set of points.

convex_mesh_area
The area of a convex mesh.

convex_mesh_mass_properties
The mass properties of a convex mesh.

convex_mesh_volume_and_center_of_mass
The volume and center of mass of a convex mesh.

convex_polyline_area
The area of a convex polyline.

convex_polyline_area_and_center_of_mass
The area and center of mass of a 2D convex Polyline.

convex_polyline_mass_properties
The mass properties of a 2D convex Polyline.

cuboid_area
The area of a cuboid.

cuboid_center_of_mass
The center of mass of a cuboid.

cuboid_unit_angular_inertia
The unit angular inertia of a cuboid.

cuboid_volume
The volume of a cuboid.

cylinder_area
The area of a cylinder.

cylinder_center_of_mass
The center of mass of a cylinder.

cylinder_unit_angular_inertia
The unit angular inertia of a cylinder.

cylinder_volume
The volume of a cylinder.