Trait nphysics3d::volumetric::InertiaTensor

pub trait InertiaTensor<N, P, AV, M> {
    fn apply(&self, a: &AV) -> AV;
    fn to_world_space(&self, &M) -> Self;
    fn to_relative_wrt_point(&self, N, &P) -> Self;
}

Trait implemented by inertia tensors.

Required Methods

fn apply(&self, a: &AV) -> AV

Applies this inertia tensor to a vector.

This is usually done by a matrix-vector multiplication.

fn to_world_space(&self, &M) -> Self

Transforms this inertia tensor from local space to world space.

fn to_relative_wrt_point(&self, N, &P) -> Self

Computes this inertia tensor relative to a given point.