Struct nphysics3d::object::ColliderData

pub struct ColliderData<N: Real> { /* fields omitted */ }

Data stored into each collider.

Those are needed by nphysics.

Methods

impl<N: Real> ColliderData<N>

pub fn new(
margin: N,
body: BodyHandle,
body_status_dependent_ndofs: usize,
position_wrt_body: Isometry<N>,
material: Material<N>
) -> Self

Initializes data for a collider.

pub fn body(&self) -> BodyHandle

The handle of the body part this collider is attached to.

pub fn margin(&self) -> N

The collision margin surrounding this collider.

pub fn position_wrt_body(&self) -> &Isometry<N>

The position of the collider relative to the body it is attached to.

pub fn material(&self) -> &Material<N>

The material of this collider.