Struct nphysics3d::solver::IntegrationParameters

pub struct IntegrationParameters<N: Real> {
    pub dt: N,
    pub erp: N,
    pub warmstart_coeff: N,
    pub restitution_velocity_threshold: N,
    pub allowed_linear_error: N,
    pub allowed_angular_error: N,
    pub max_linear_correction: N,
    pub max_angular_correction: N,
    pub max_stabilization_multiplier: N,
    pub max_velocity_iterations: usize,
    pub max_position_iterations: usize,
}

Parameters for a time-step of the physics engine.

Fields
dt: N

The timestep (default: 1.0 / 60.0)

erp: N

The Error Reduction Parameter in [0, 1] is the proportion of the positional error to be corrected at each time step (default: 0.2).

warmstart_coeff: N

Each cached impulse are multiplied by this coefficient in [0, 1] when they are re-used to initialize the solver (default 1.0).

restitution_velocity_threshold: N

Contacts at points where the involved bodies have a relative velocity smaller than this threshold wont be affected by the restitution force (default: 1.0).

allowed_linear_error: N

Ammount of penetration the engine wont attempt to correct (default: 0.001m).

allowed_angular_error: N

Ammount of angular drift of joint limits the engine wont attempt to correct (default: 0.001rad).

max_linear_correction: N

Maximum linear correction during one step of the non-linear position solver (default: 100.0).

max_angular_correction: N

Maximum angular correction during one step of the non-linear position solver (default: 0.2).

max_stabilization_multiplier: N

Maximum nonlinera SOR-prox scaling parameter when the constraint correction direction is close to the kernel of the involved multibody's jacobian (default: 0.2).

max_velocity_iterations: usize

Maximum number of iterations performed by the velocity constraints solver.

max_position_iterations: usize

Maximum number of iterations performed by the position-based constraints solver.

Methods

impl<N: Real> IntegrationParameters<N>

pub fn new(
dt: N,
erp: N,
warmstart_coeff: N,
restitution_velocity_threshold: N,
allowed_linear_error: N,
allowed_angular_error: N,
max_linear_correction: N,
max_angular_correction: N,
max_stabilization_multiplier: N,
max_velocity_iterations: usize,
max_position_iterations: usize
) -> Self

Creates a set of integration parameters with the given values.